Replaces Material.Icon with Material.Icons
Material.Icon which is build into elm-mdl provides a nice and
easy way to get material icons with a default style (color and size)
with very little effort.
However since controlling the color per icon is one of our goals
we decided to switch to material-icons which provides Material.Icons.
However the generated functions expect a color and a size argument,
which led to some confusion because they are expecting Color.Color
and not Material.Color.Color which was hard to understand, because
we aliased Material.Color to Color which caused type errors like
```
expected: Color.Color
got:      Color.Color
```
To fix this we changed the material color module alias to MColor.
This changeset includes the new aliases and changes to the places
that already used it.