Further implementation of body index calculations
Adds a new module that provides all the body index calculations.
In addition most of the static content has been switched to use
real calculations.
Currently we use Result.withDefault but in the future those functions
will also return a `Result String Float` and we will combine to
Result processing with the toString conversion to convert missing
results into N/A instead of -1.