Switch number input validation to Regex
The generative tests showed that currently "+" and "-" lead
to wrong results (Ok NaN) which is not intended.
After talking to some community members this seems to be known and
might be changed in elm 0.19. However for now using Regex instead
seems to be the most reasonable way to continue.
(slack discussion link: https://elmlang.slack.com/archives/C0CJ3SBBM/p1499517209549234)