Improves styling and cleans up navi
In order to improve the overall design consistency the body index
and body fat views have been made similar with regards to padding
and spacing. In addition the custom grid for the gender selection
has been dropped.
As a final measure for this iteration we replaced the fake nav
entries with proper links.