Implements fully functional bash based navigation
As the first step this change implmements a fully functional
url-hash based SPA navigation. The primary reason for this is
that we do not require to set up a backend up for handling
different URLs and passing them to the FE app.
All other parts would work fine with preventDefaults etc, but
this seems a big drawback, since I consider bringing FE and BE
closer together and I want to avoid any decision concering that
for now.
There are also some redudant and obsolete functions for the navigation
at the moment as well as some fixmes. They will be addresses soon.