Implement translations and improve styling
Adds I18n helper to replace static text strings with localized
keys. The implementation is heavily inspired by this
[blog post](https://www.gizra.com/content/elm-i18n-type-safety/).

In addition to that styles for screens with less than 400px have been
improved.

In addition a unsupported test case has been fixed. The fuzzer
noted that 0.0 as input is not properly checked.