Adds hacky workarounds to fix tab bar styles.
I guess there are better and correct ways to do this, but the lack
of documentation and not not working instructions in code comments
did not help. Thus two simple hacks have been added to ensure that
we do not have a scrolling navigation and that the brokenly styled
section is not shown. (width seems to be computed in the wrong way.