Improves welcome content grid
We had an issue with empty space between cards on tablet due to
our hard coded 3 cards per row design. We switched to a simpler
n in a row model and now the grid module can take care of the
row layout.
In addition we improved the english translations for the current
dummy cards.