Make input optional and remove defaults
In order to remove all default values and to actually allow having
not filled in information without rendering errors, as well as having
initial empty state as valid, we needed a more complex type for
representing input: OptionalValidatedInput = Maybe Result String Float.
However we realized that we can use more narrow types in our calculation
by converting this to a Maybe Float. We don't care about input errors
in our computation and thus we just simplified the type before passing
it to the body-index and body-fat logic. This required only a simple
helper function and 2 additional type alises.
We now distinguish between Input records and value records and
convert between them.
This made the transition quiet easy and also separated the concerns
of input validation / error presentation from computation.