Removes labels from validations.
As the second to last step we removed the concept of labels/prefix
for validation functions since we can simply use mapError on the
call site. This reduces coupling and should make translations easier.
However we will skip localizing the validation error until I have
a better understanding on whether I should pass a locale to the
validators or return a I18n.Key which conflices with the result of
parseFloat.