Implement insights
In order to get a proper understanding of the customer behaviour
the following measures have been implemented.
* Google analytics page view tracking
* Google analytics SPA page view tracking for clicks
* Google analytics event tracking for form submissions
* Form submit backup / persistence to old backend

This should help to get a proper understanding of who the customers
are and what they want to do on the page. This should be a valuable
asset for deciding which features to implement next.