Replaces tabs with adaptive drawer.
Instead of using tabs and a drawer I decided to remove the
tabbar for now. According to the material UI guides it is also
acceptable to to have a persistent drawer on large screen and a
normal one for small screen. However on tablet it should be
closable which does not seem to be implemented in elm mdl. Thus
we took a close appraoch. Spec can be found here:
https://material.io/guidelines/patterns/navigation.html#navigation-patterns

In addition the fact that we are going to add at least 2-3 more items
also makes the drawer a better fit according to the guides.